# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  overdue: "overdue",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type CardVisualStatus =
  (typeof CardVisualStatus)[keyof typeof CardVisualStatus];

export const CardVisualStatus = {
  no_task: "no_task",
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  overdue: "overdue",
} as const;

export type WorkspaceRole = (typeof WorkspaceRole)[keyof typeof WorkspaceRole];

export const WorkspaceRole = {
  admin: "admin",
  editor: "editor",
  executor: "executor",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserResponse {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: UserResponse;
  token: string;
}

export interface CreateWorkspaceRequest {
  name: string;
}

export interface WorkspaceResponse {
  id: string;
  name: string;
  createdBy: string;
  createdAt: string;
  role: WorkspaceRole;
}

export interface WorkspaceMemberResponse {
  id: string;
  workspaceId: string;
  userId: string;
  role: WorkspaceRole;
  user: UserResponse;
  createdAt: string;
}

export interface MapResponse {
  id: string;
  workspaceId: string;
  name: string;
  createdBy: string;
  createdAt: string;
  updatedAt: string;
}

export interface WorkspaceDetailResponse {
  id: string;
  name: string;
  createdBy: string;
  createdAt: string;
  role: WorkspaceRole;
  members: WorkspaceMemberResponse[];
  maps: MapResponse[];
}

export interface AddMemberRequest {
  email: string;
  role: WorkspaceRole;
}

export interface UpdateMemberRoleRequest {
  role: WorkspaceRole;
}

export interface CreateMapRequest {
  name: string;
}

export interface CardResponse {
  id: string;
  mapId: string;
  title: string;
  description?: string | null;
  positionX: number;
  positionY: number;
  statusVisual: CardVisualStatus;
  taskId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConnectionResponse {
  id: string;
  mapId: string;
  sourceCardId: string;
  targetCardId: string;
  createdAt: string;
}

export interface MapDetailResponse {
  id: string;
  workspaceId: string;
  name: string;
  createdBy: string;
  createdAt: string;
  updatedAt: string;
  cards: CardResponse[];
  connections: ConnectionResponse[];
}

export interface CreateCardRequest {
  title: string;
  description?: string;
  positionX?: number;
  positionY?: number;
}

export interface UpdateCardRequest {
  title?: string;
  description?: string;
  positionX?: number;
  positionY?: number;
}

export interface TaskResponse {
  id: string;
  mapId: string;
  workspaceId: string;
  title: string;
  description?: string | null;
  assignedTo?: string | null;
  assignedUser?: UserResponse | null;
  dueDate?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CardDetailResponse {
  id: string;
  mapId: string;
  title: string;
  description?: string | null;
  positionX: number;
  positionY: number;
  statusVisual: CardVisualStatus;
  taskId?: string | null;
  task?: TaskResponse | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConnectionRequest {
  sourceCardId: string;
  targetCardId: string;
}

export interface CreateTaskRequest {
  title: string;
  description?: string;
  assignedTo?: string | null;
  dueDate?: string | null;
  priority?: TaskPriority;
}

export interface UpdateTaskStatusRequest {
  status: TaskStatus;
}

export interface UpdateTaskDetailsRequest {
  title?: string;
  description?: string;
  assignedTo?: string | null;
  dueDate?: string | null;
  priority?: TaskPriority;
}

export interface TaskWithContextResponse {
  id: string;
  mapId: string;
  workspaceId: string;
  title: string;
  description?: string | null;
  assignedTo?: string | null;
  dueDate?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  cardId: string;
  cardTitle: string;
  mapName: string;
  workspaceName: string;
}

export type DashboardResponseTasksByStatus = {
  pending: number;
  in_progress: number;
  completed: number;
  overdue: number;
};

export type DashboardResponseTasksByPriority = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface DashboardResponse {
  workspaceId: string;
  totalMaps: number;
  totalCards: number;
  totalTasks: number;
  tasksByStatus: DashboardResponseTasksByStatus;
  tasksByPriority: DashboardResponseTasksByPriority;
}

export type GetMyTasksParams = {
  workspaceId?: string;
  status?: TaskStatus;
};