# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MindTask API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workspaces/{workspaceId}/maps/{mapId}/cards:
    post:
      operationId: createCard
      tags: [cards]
      summary: Create a card in the map
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
  /workspaces/{workspaceId}/maps/{mapId}/cards/{cardId}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card with its task
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Card with optional task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetailResponse"
    put:
      operationId: updateCard
      tags: [cards]
      summary: Update a card (title, description, position)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete a card
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /workspaces/{workspaceId}/maps/{mapId}/cards/{cardId}/task:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task linked to a card
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created and card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "409":
          description: Card already has a task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unlinkTask
      tags: [tasks]
      summary: Unlink and delete task from a card
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task unlinked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/maps/{mapId}/cards/{cardId}/task/status:
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      summary: Update task status (syncs card visual status)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStatusRequest"
      responses:
        "200":
          description: Task and card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /workspaces/{workspaceId}/maps/{mapId}/cards/{cardId}/task/details:
    patch:
      operationId: updateTaskDetails
      tags: [tasks]
      summary: Update task details (title, description, assignee, due date, priority)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskDetailsRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed, overdue]
    TaskPriority:
      type: string
      enum: [low, medium, high, critical]
    CardVisualStatus:
      type: string
      enum: [no_task, pending, in_progress, completed, overdue]
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        token:
          type: string
      required: [user, token]
...
    CreateCardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        positionX:
          type: number
        positionY:
          type: number
      required: [title]
    UpdateCardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        positionX:
          type: number
        positionY:
          type: number
    CardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        positionX:
          type: number
        positionY:
          type: number
        statusVisual:
          $ref: "#/components/schemas/CardVisualStatus"
        taskId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mapId, title, positionX, positionY, statusVisual, createdAt, updatedAt]
    CardDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        positionX:
          type: number
        positionY:
          type: number
        statusVisual:
          $ref: "#/components/schemas/CardVisualStatus"
        taskId:
          type: string
          format: uuid
          nullable: true
        task:
          $ref: "#/components/schemas/TaskResponse"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mapId, title, positionX, positionY, statusVisual, createdAt, updatedAt]
...
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: string
          format: uuid
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          $ref: "#/components/schemas/TaskPriority"
      required: [title]
    UpdateTaskStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/TaskStatus"
      required: [status]
    UpdateTaskDetailsRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedTo:
          type: string
          format: uuid
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          $ref: "#/components/schemas/TaskPriority"
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        assignedTo:
          type: string
          format: uuid
          nullable: true
        assignedUser:
          $ref: "#/components/schemas/UserResponse"
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mapId, workspaceId, title, priority, status, createdAt, updatedAt]
...