# lib/api-zod/src/generated/types/cardVisualStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */

export type CardVisualStatus =
  (typeof CardVisualStatus)[keyof typeof CardVisualStatus];

export const CardVisualStatus = {
  no_task: "no_task",
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  overdue: "overdue",
} as const;