# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserResponse {
  id: string;
  name: string;
  email: string;
  createdAt: Date;
}